+++ sdk/pixFraud/index.js
@@ -1,0 +1,9 @@
+const pixFraud = require('./pixFraud.js');
+
+exports.log = require('./log');
+exports.create = pixFraud.create;
+exports.get = pixFraud.get;
+exports.query = pixFraud.query;
+exports.page = pixFraud.page;
+exports.cancel = pixFraud.cancel;
+exports.PixFraud = pixFraud.PixFraud;