+++ sdk/pixKeyHolmes/index.js
@@ -1,0 +1,6 @@
+const pixKeyHolmes = require('./pixKeyHolmes.js');
+
+exports.create = pixKeyHolmes.create;
+exports.query = pixKeyHolmes.query;
+exports.page = pixKeyHolmes.page;
+exports.PixKeyHolmes = pixKeyHolmes.PixKeyHolmes;